feat: add ExitMode enum for process exit handling (#25)

+++ src/main.rs
@@ -8,5 +8,5 @@ fn main() {
     Builder::from_env(Env::default().default_filter_or("info"))
         .target(Target::Stdout)
         .init();
-    process::exit(compute::app_runner::start());
+    process::exit(compute::app_runner::start() as i32);
 }
